Separated the list declaration/definition from the translator.
